requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
click==8.0.1; python_version >= '3.6'
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0; python_version >= '3.6'
pydantic==1.8.2
starlette==0.14.2; python_version >= '3.6'
typing-extensions==3.10.0.2
uvicorn==0.15.0
tortoise-orm
requests
asyncpg
celery
redis

